requirements.txt
click==6.7
RPi.GPIO==0.6.3

